# openapi.yaml
openapi: 3.1.0
info:
  title: Crypto Vision API
  description: >
    The complete cryptocurrency intelligence API. Aggregates data from CoinGecko,
    DeFiLlama, DexScreener, RSS feeds, and LLM providers into a single,
    high-performance endpoint.
  version: 0.1.0
  contact:
    name: Crypto Vision
    url: https://cryptocurrency.cv
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://cryptocurrency.cv
    description: Production
...
paths:
...
  /api/coin/{id}:
    get:
      operationId: getCoinDetail
      summary: Coin detail
      description: Returns full detail for a single coin including description, links, categories, and market data.
      tags: [Market]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: CoinGecko coin ID (e.g. bitcoin, ethereum)
      responses:
        "200":
          description: Coin detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/CoinDetail"
                  timestamp:
                    type: string
                    format: date-time
  /api/price:
    get:
      operationId: getPrice
      summary: Simple price lookup
      description: Returns current prices for one or more coins in the requested fiat currencies.
      tags: [Market]
      parameters:
        - name: ids
          in: query
          required: true
          schema:
            type: string
          description: Comma-separated CoinGecko coin IDs (e.g. bitcoin,ethereum)
        - name: vs_currencies
          in: query
          schema:
            type: string
            default: usd
          description: Comma-separated target currencies (e.g. usd,eur)
      responses:
        "200":
          description: Price map
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    additionalProperties:
                      type: object
                      additionalProperties:
                        type: number
                  timestamp:
                    type: string
                    format: date-time
        "400":
          description: Missing ids parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/search:
    get:
      operationId: searchCoins
      summary: Search coins
      description: Search for coins by name or symbol.
      tags: [Market]
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        symbol:
                          type: string
                        rank:
                          type: integer
                          nullable: true
                  timestamp:
                    type: string
                    format: date-time
        "400":
          description: Missing q parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ──────────────────────────────────────────────────────────────
  # Semantic Search
  # ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Search ─────────────────────────────────────────────────
...
    CoinDetail:
      type: object
      properties:
        id:
          type: string
        symbol:
          type: string
        name:
          type: string
        description:
          type: string
          description: English description
        categories:
          type: array
          items:
            type: string
        platforms:
          type: object
          additionalProperties:
            type: string
        links:
          type: object
          properties:
            homepage:
              type: array
              items:
                type: string
                format: uri
            explorers:
              type: array
              items:
                type: string
                format: uri
            github:
              type: array
              items:
                type: string
                format: uri
        marketData:
          type: object
          properties:
            price:
              type: object
              additionalProperties:
                type: number
            marketCap:
              type: object
              additionalProperties:
                type: number
            volume:
              type: object
              additionalProperties:
                type: number
            change24h:
              type: number
              nullable: true
            change7d:
              type: number
              nullable: true
            change30d:
              type: number
              nullable: true
            circulatingSupply:
              type: number
            totalSupply:
              type: number
              nullable: true
            maxSupply:
              type: number
              nullable: true
...